virtual strip show method optimization
the virtual strip "show" method iterated through every pixel of the
strip to construct a list of strips.  Pretty gross since you are calling
it every frame, like 30, 60, 100 times a second.  Since the list of
physical strips can only change when new physical LEDs are added to the
virtual strip, we manage the set in add_pixel_range, which is the only
place where the class can be given new physical strips.